- database finished

+++ app/database/migrations/2020_04_17_114154_create_images_table.php
@@ -14,9 +14,10 @@ class CreateImagesTable extends Migration
     public function up()
     {
         Schema::create('images', function (Blueprint $table) {
-            $table->increments('id');
-            $table->unsignedInteger('user_id');
+            $table->bigIncrements('id');
+            $table->unsignedBigInteger('user_id');
             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
+
             $table->string('title')->nullable();
             $table->string('path')->unique();
             $table->text('description')->nullable();
